Store a label's associated heading
The end goal is to remove the heading ctags altogether. The logic to
find a label's associated heading using ctags is a bit more complex than
I'd like. Storing headings as taglines also takes up space in the ctags file.

This commit attempts to solve these issues by moving the heading search
logic into the markdown parsing step and storing only the associated
heading (not all headings) as a tagfield (not a whole tagline).

We still rely on headings to find entry files without labels. We'll
tackle that in a separate commit.