Handle entries with no labels or headings
I'm not in love with this solution - it just doesn't look elegant to
me - but it will do for now.

Rather than converting everything to a list of taglines, perhaps we
should be iterating over files and then look up the label/title metadata
in the taglines. Maybe some kind of (sortable) hashmap of filename to
tags.

Now that we've removed headings, we don't necessarily need to keep a
pointer associating label occurences with a linked list of tags
(previously used to look up a label's heading).